+++ CHANGELOG.md
@@ -5,9 +5,9 @@ the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
 ## Unreleased
-- Fixed: Issue where isolating asset without pivot data would break.
-
--   Added: ExternalFiles Youtube plugin
+-   Added: ExternalFiles Youtube plugin.
+-   Fixed: Issue where isolating asset without pivot data would break.
+-   Fixed: Internal pages multiselect in menu item edit didn't show results.
 
 ## 0.8.4 - 2023-10-30