+++ CHANGELOG.md
@@ -5,6 +5,8 @@ the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
 ## Unreleased
+
+## 0.8.5 - 2023-10-30
 -   Added: ExternalFiles Youtube plugin.
 -   Fixed: Issue where isolating asset without pivot data would break.
 -   Fixed: Internal pages multiselect in menu item edit didn't show results.